Code cleanup
 - Added a 10MB limit for the kernel, for now. This smaller limit will
have me try and save space and not assume a ton of memory.

 - Fixed the gdt assembly by using proper comments with a semicolor.

 - Renamed the gdt assembly file to agdt.asm (agdt for assembly gdt)
because it was called gdt.s and the gdt.c compilation was overwriting
the gdt.o that was being produced in both cases.

 - Fixed compiler warnings by setting proper types explicitely.

+++ gdt.c
@@ -16,7 +16,7 @@ void gdtSetGate(int num, uint64_t base, uint64_t limit, uint8_t access, uint8_t 
 
 void setupGdt(){
    gdtPointer.limit   = (sizeof(gdtEntry_t) * 3) - 1;
-   gdtPointer.base      = &gdt;
+   gdtPointer.base      = (uint32_t)&gdt;
 
    gdtSetGate(0, 0, 0, 0, 0);
    gdtSetGate(1, 0, 0xFFFFFFFF, 0x9A, 0xCF);
